remove the testlog package
