+++ integrationtests/self/self_suite_test.go
@@ -1,15 +1,20 @@
 package self_test
 
 import (
+	"bytes"
 	"crypto/tls"
+	"flag"
+	"log"
 	"math/rand"
+	"os"
+	"sync"
 	"testing"
+
+	"github.com/lucas-clemente/quic-go/internal/testdata"
+	"github.com/lucas-clemente/quic-go/internal/utils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	_ "github.com/lucas-clemente/quic-go/integrationtests/tools/testlog"
-	"github.com/lucas-clemente/quic-go/internal/testdata"
 )
 
 const alpn = "quic-go integration tests"
@@ -48,6 +53,73 @@ func GeneratePRData(l int) []byte {
 		res[i] = byte(seed)
 	}
 	return res
+}
+
+const logBufSize = 100 * 1 << 20 // initial size of the log buffer: 100 MB
+
+type syncedBuffer struct {
+	mutex sync.Mutex
+
+	*bytes.Buffer
+}
+
+func (b *syncedBuffer) Write(p []byte) (int, error) {
+	b.mutex.Lock()
+	n, err := b.Buffer.Write(p)
+	b.mutex.Unlock()
+	return n, err
+}
+
+func (b *syncedBuffer) Bytes() []byte {
+	b.mutex.Lock()
+	p := b.Buffer.Bytes()
+	b.mutex.Unlock()
+	return p
+}
+
+func (b *syncedBuffer) Reset() {
+	b.mutex.Lock()
+	b.Buffer.Reset()
+	b.mutex.Unlock()
+}
+
+var (
+	logFileName string // the log file set in the ginkgo flags
+	logBufOnce  sync.Once
+	logBuf      *syncedBuffer
+)
+
+// read the logfile command line flag
+// to set call ginkgo -- -logfile=log.txt
+func init() {
+	flag.StringVar(&logFileName, "logfile", "", "log file")
+}
+
+var _ = BeforeEach(func() {
+	log.SetFlags(log.Ldate | log.Ltime | log.Lmicroseconds)
+
+	if debugLog() {
+		logBufOnce.Do(func() {
+			logBuf = &syncedBuffer{Buffer: bytes.NewBuffer(make([]byte, 0, logBufSize))}
+		})
+		utils.DefaultLogger.SetLogLevel(utils.LogLevelDebug)
+		log.SetOutput(logBuf)
+	}
+})
+
+var _ = AfterEach(func() {
+	if debugLog() {
+		logFile, err := os.Create(logFileName)
+		Expect(err).ToNot(HaveOccurred())
+		logFile.Write(logBuf.Bytes())
+		logFile.Close()
+		logBuf.Reset()
+	}
+})
+
+// Debug says if this test is being logged
+func debugLog() bool {
+	return len(logFileName) > 0
 }
 
 func TestSelf(t *testing.T) {
