--- integrationtests/tools/testlog/testlog.go
@@ -1,83 +1,0 @@
-package testlog
-
-import (
-	"bytes"
-	"flag"
-	"log"
-	"os"
-	"sync"
-
-	"github.com/lucas-clemente/quic-go/internal/utils"
-
-	//nolint:stylecheck
-	. "github.com/onsi/ginkgo"
-	//nolint:stylecheck
-	. "github.com/onsi/gomega"
-)
-
-const logBufSize = 100 * 1 << 20 // initial size of the log buffer: 100 MB
-
-type syncedBuffer struct {
-	mutex sync.Mutex
-
-	*bytes.Buffer
-}
-
-func (b *syncedBuffer) Write(p []byte) (int, error) {
-	b.mutex.Lock()
-	n, err := b.Buffer.Write(p)
-	b.mutex.Unlock()
-	return n, err
-}
-
-func (b *syncedBuffer) Bytes() []byte {
-	b.mutex.Lock()
-	p := b.Buffer.Bytes()
-	b.mutex.Unlock()
-	return p
-}
-
-func (b *syncedBuffer) Reset() {
-	b.mutex.Lock()
-	b.Buffer.Reset()
-	b.mutex.Unlock()
-}
-
-var (
-	logFileName string // the log file set in the ginkgo flags
-	logBufOnce  sync.Once
-	logBuf      *syncedBuffer
-)
-
-// read the logfile command line flag
-// to set call ginkgo -- -logfile=log.txt
-func init() {
-	flag.StringVar(&logFileName, "logfile", "", "log file")
-}
-
-var _ = BeforeEach(func() {
-	log.SetFlags(log.Ldate | log.Ltime | log.Lmicroseconds)
-
-	if Debug() {
-		logBufOnce.Do(func() {
-			logBuf = &syncedBuffer{Buffer: bytes.NewBuffer(make([]byte, 0, logBufSize))}
-		})
-		utils.DefaultLogger.SetLogLevel(utils.LogLevelDebug)
-		log.SetOutput(logBuf)
-	}
-})
-
-var _ = AfterEach(func() {
-	if Debug() {
-		logFile, err := os.Create(logFileName)
-		Expect(err).ToNot(HaveOccurred())
-		logFile.Write(logBuf.Bytes())
-		logFile.Close()
-		logBuf.Reset()
-	}
-})
-
-// Debug says if this test is being logged
-func Debug() bool {
-	return len(logFileName) > 0
-}
